pyproject.toml
[tool.poetry]
name = "Behavox"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv"], version = "1.9.1"}
python-multipart = "^0.0.5"
requests = "^2.27.1"
setproctitle = "^1.2.3"
starlette-prometheus = "^0.9.0"
uvicorn = "^0.17.6"
PyYAML = "^6.0"
alembic = "^1.7.7"
SQLAlchemy = "^1.4.36"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.25.0"
python-json-logger = "^2.0.2"
rfc5424-logging-handler = "^1.4.3"
celery = "^5.1.2"
flower = "^1.0.0"
dill = "0.3.4"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.3.0"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.4.25"
flake8-coding = "^1.3.2"
flake8-comprehensions = "^3.10.0"
flake8-debugger = "^4.1.2"
flake8-eradicate = "^1.2.1"
flake8-executable = "^2.1.1"
flake8-isort = "^4.1.1"
flake8-logging-format = "^0.6.0"
flake8-pie = "^0.15.0"
flake8-pytest-style = "^1.6.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
pytest-xdist = "^2.5.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-env = "^0.6.2"
pytest-aiohttp = "^1.0.4"


[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true

[tool.isort]
line_length = 120
combine_as_imports = true
order_by_type = true
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
known_third_party = ["alembic"]

[tool.coverage.run]
branch = true
omit = ["*tests*", "*venv*", "alembic/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
]

[tool.pytest.ini_options]
addopts = "--failed-first --exitfirst --showlocals"
python_files = "tests.py test_*.py"
python_functions = "test_*"
env = [
    "WEB_CONCURRENCY=1",

    "DB_POOL_SIZE=5",
    "DB_MAX_OVERFLOW=5",
    "DB_ECHO=False",
    "DB_URL=postgresql+asyncpg://postgres:postgres@web_db/test_db",

    "ALEMBIC_MIGRATION_VERSION_TABLE=scheduler_alembic_migration_version",

    "WORKERS_NAME = [\"py36\",\"py38\"]",

    "CELERY_BROKER_URL=pyamqp://scheduler:scheduler@rabbitmq:5672/scheduler",
    "CELERY_RESULT_BACKEND=db+postgresql://postgres:postgres@web_db:5432/main_db"
  ]
